Move connect_to_db_or_raise(), exceptions to "common"
