+++ apps/podcast-transcribe-episode/bin/rabbitmq_worker.py
@@ -4,13 +4,13 @@ import asyncio
 
 from mediawords.job import JobBroker
 from mediawords.util.log import create_logger
-from mediawords.util.network import wait_for_tcp_port_to_open
 from mediawords.util.perl import decode_object_from_bytes_if_needed
+from mediawords.workflow.client import workflow_client
 
 # noinspection PyPackageRequirements
 from temporal.workflow import WorkflowClient, WorkflowOptions
 
-from podcast_transcribe_episode.workflow_interface import NAMESPACE, AbstractPodcastTranscribeWorkflow
+from podcast_transcribe_episode.workflow_interface import AbstractPodcastTranscribeWorkflow
 
 log = create_logger(__name__)
 
@@ -18,10 +18,7 @@ log = create_logger(__name__)
 async def _start_workflow(stories_id: int) -> None:
     log.info(f"Starting a workflow for story {stories_id}...")
 
-    # FIXME it's super lame to wait for this port to open, but the Python SDK seems to fail otherwise
-    wait_for_tcp_port_to_open(hostname='temporal-server', port=7233)
-
-    client = WorkflowClient.new_client(host='temporal-server', namespace=NAMESPACE)
+    client = workflow_client()
     workflow: AbstractPodcastTranscribeWorkflow = client.new_workflow_stub(
         cls=AbstractPodcastTranscribeWorkflow,
         workflow_options=WorkflowOptions(workflow_id=str(stories_id)),
