+++ apps/podcast-transcribe-episode/src/python/podcast_transcribe_episode/media_info.py
@@ -7,8 +7,8 @@ from typing import Type, Optional, List, Any, Dict
 import ffmpeg
 
 from mediawords.util.log import create_logger
+from mediawords.workflow.exceptions import McProgrammingError, McPermanentError
 
-from .exceptions import McProgrammingError, McPermanentError
 from .audio_codecs import (
     AbstractAudioCodec,
     Linear16AudioCodec,
